services/user/requirements.txt
celery==4.4.6
Django==3.0.8
django-cors-headers==3.4.0
djangogrpcframework==0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
grpcio==1.27.1
grpcio-tools==1.27.1
gunicorn==20.0.4
Pillow==7.2.0
protobuf==3.12.2
psycopg2-binary==2.8.5
python-dateutil==2.8.1
redis==3.5.3
